boardController 분리
